+++ doc/source/changelog.rst
@@ -2,6 +2,14 @@
 Change Log
 ==========
 
+v1.1.0
+====================
+
+* Fixed `Rename render_static -> renderstatic <https://github.com/bckohan/django-render-static/issues/11>`_
+* Fixed `Richer contexts <https://github.com/bckohan/django-render-static/issues/6>`_
+* Fixed `Add GET query parameters to ClassURLWriter's reverse function <https://github.com/bckohan/django-render-static/issues/12>`_
+
+
 v1.0.1
 ====================